requirements.txt
flask
flask-cors
requests
python-whois
dnspython
openpyxl
beautifulsoup4
